Implement load tests.
The tests all pass in normal mode, and all fail for CHURN.
